Moved express table integration tests
