+++ test/express/integration/tables/softDelete.tests.js
@@ -4,9 +4,9 @@
 var expect = require('chai').use(require('chai-subset')).expect,
     supertest = require('supertest-as-promised'),
     express = require('express'),
-    mobileApps = require('../infrastructure/mobileApps'),
-    data = require('../../../src/data/mssql'),
-    config = require('../infrastructure/config'),
+    mobileApps = require('../../infrastructure/mobileApps'),
+    data = require('../../../../src/data/mssql'),
+    config = require('../../infrastructure/config'),
 
     app, mobileApp;
